using node thread pool for file export
