+++ src/datachain/lib/dc.py
@@ -23,6 +23,7 @@ import sqlalchemy
 from pydantic import BaseModel
 from sqlalchemy.sql.functions import GenericFunction
 from sqlalchemy.sql.sqltypes import NullType
+from tqdm import tqdm
 
 from datachain.dataset import DatasetRecord
 from datachain.func import literal
@@ -32,7 +33,14 @@ from datachain.lib.convert.python_to_sql import python_to_sql
 from datachain.lib.convert.values_to_tuples import values_to_tuples
 from datachain.lib.data_model import DataModel, DataType, DataValue, dict_to_data_model
 from datachain.lib.dataset_info import DatasetInfo
-from datachain.lib.file import ArrowRow, File, FileType, get_file_type
+from datachain.lib.file import (
+    EXPORT_FILES_MAX_THREADS,
+    ArrowRow,
+    File,
+    FileExporter,
+    FileType,
+    get_file_type,
+)
 from datachain.lib.file import ExportPlacement as FileExportPlacement
 from datachain.lib.listing import get_file_info, get_listing, list_bucket, ls
 from datachain.lib.listing_info import ListingInfo
@@ -67,54 +75,6 @@ UDFObjT = TypeVar("UDFObjT", bound=UDFBase)
 
 
 DEFAULT_PARQUET_CHUNK_SIZE = 100_000
-
-from datachain.nodes_thread_pool import NodesThreadPool
-from tqdm import tqdm
-
-
-class FileExporter(NodesThreadPool):
-    def __init__(
-        self,
-        output: str,
-        placement: FileExportPlacement,
-        use_cache: bool,
-        link_type: Literal["copy", "symlink"],
-        num_files: int,
-        max_threads: int = 1,
-    ):
-        super().__init__(max_threads)
-        '''
-        self.progress_bar = tqdm(
-            desc=f"Exporting files to {output}: ",
-            unit=" files",
-            unit_scale=True,
-            unit_divisor=1000,
-            total=num_files,  # type: ignore [union-attr]
-            leave=False,
-        )
-        '''
-        self.output = output
-        self.placement = placement
-        self.use_cache = use_cache
-        self.link_type = link_type
-
-    def done_task(self, done):
-        for task in done:
-            task.result()
-
-    def do_task(self, file):
-        file.export(
-            self.output, self.placement, self.use_cache, link_type=self.link_type
-        )  # type: ignore[union-attr]
-        '''
-        for file in files:
-            file.export(
-                self.output, self.placement, self.use_cache, link_type=self.link_type
-            )  # type: ignore[union-attr]
-        '''
-        # self.increase_counter(1)  # type: ignore [arg-type]
-        # sometimes progress bar doesn't get updated so manually updating it
-        # self.update_progress_bar(self.progress_bar)
 
 
 def resolve_columns(
@@ -2530,7 +2490,7 @@ class DataChain:
         placement: FileExportPlacement = "fullpath",
         use_cache: bool = True,
         link_type: Literal["copy", "symlink"] = "copy",
-        parallel: Optional[int] = None,
+        num_workers: Optional[int] = EXPORT_FILES_MAX_THREADS,
     ) -> None:
         """Export files from a specified signal to a directory. Files can be
         exported to a local or cloud directory.
@@ -2546,9 +2506,8 @@ class DataChain:
             prefetch: number of workers to use for downloading files in advance.
                       This is enabled by default and uses 2 workers.
                       To disable prefetching, set it to 0.
-            parallel : number of workers for exporting in parallel.
-                By default it uses 2 workers.
-                To disable parallel exporting, set it to 0.
+            num_workers : number of workers to use for exporting files.
+                By default it uses 5 workers.
 
         Example:
             Cross cloud transfer
@@ -2576,26 +2535,9 @@ class DataChain:
             placement,
             use_cache,
             link_type,
-            self.count(),
+            max_threads=num_workers or 1,
         )
-        try:
-            file_exporter.run(
-                self.collect(signal), progress_bar
-            )
-        except:
-            print("Error while exporting files")
-            raise
-        '''
-        if parallel == 0:
-            for file in self.collect(signal):
-                file.export(output, placement, use_cache, link_type=link_type)  # type: ignore[union-attr]
-        self.settings(cache=use_cache, prefetch=parallel).map(
-            res=lambda file: file.export(
-                output, placement, use_cache, link_type=link_type
-            ),
-            params=[signal],
-        ).exec()
-        '''
+        file_exporter.run(self.collect(signal), progress_bar)
 
     def shuffle(self) -> "Self":
         """Shuffle the rows of the chain deterministically."""
